+++ TabletopTweaks-Base/Bugfixes/Items/Armor.cs
@@ -1,8 +1,10 @@
 ﻿using HarmonyLib;
+using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Items.Armors;
 using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.Items;
 using Kingmaker.UnitLogic;
+using Kingmaker.UnitLogic.FactLogic;
 using System;
 using TabletopTweaks.Core.Utilities;
 using static TabletopTweaks.Base.Main;
@@ -19,12 +21,21 @@ namespace TabletopTweaks.Base.Bugfixes.Items {
 
                 TTTContext.Logger.LogHeader("Patching Armor");
                 PatchHaramaki();
+                PatchSingingSteel();
 
                 void PatchHaramaki() {
                     if (TTTContext.Fixes.Items.Armor.IsDisabled("Haramaki")) { return; }
                     var HaramakiType = BlueprintTools.GetBlueprint<BlueprintArmorType>("9511d62bcfc57c245bf64350a5933470");
                     HaramakiType.m_ProficiencyGroup = ArmorProficiencyGroup.Light;
                     TTTContext.Logger.LogPatch("Patched", HaramakiType);
+                }
+                void PatchSingingSteel() {
+                    if (TTTContext.Fixes.Items.Armor.IsDisabled("SingingSteel")) { return; }
+                    var SingingSteelFeature = BlueprintTools.GetBlueprint<BlueprintFeature>("793af594fa45e6a47b89719dda7d5f7a");
+                    SingingSteelFeature.AddComponent<AddMechanicsFeature>(c => {
+                        c.m_Feature = AddMechanicsFeature.MechanicsFeatureType.SingingSteel;
+                    });
+                    TTTContext.Logger.LogPatch("Patched", SingingSteelFeature);
                 }
             }
         }